+++ src/main/java/coz/weavon/validation/model/ValidPagination.java
@@ -1,6 +1,6 @@
-package coz.weavon.common.presentation.model.validation;
+package coz.weavon.validation.model;
 
-import coz.weavon.common.presentation.validator.RangeValidator;
+import coz.weavon.validation.validator.PaginationValidator;
 import jakarta.validation.Constraint;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -9,12 +9,8 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = RangeValidator.class)
-public @interface ValidRange {
+@Constraint(validatedBy = PaginationValidator.class)
+public @interface ValidPagination {
 
     boolean required() default true;
-
-    String startFieldName() default "startDate";
-
-    String endFieldName() default "endDate";
 }
