+++ src/main/java/coz/weavon/validation/model/ValidRange.java
@@ -1,6 +1,6 @@
-package coz.weavon.common.presentation.model.validation;
+package coz.weavon.validation.model;
 
-import coz.weavon.common.presentation.validator.PaginationValidator;
+import coz.weavon.validation.validator.RangeValidator;
 import jakarta.validation.Constraint;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -9,8 +9,12 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = PaginationValidator.class)
-public @interface ValidPagination {
+@Constraint(validatedBy = RangeValidator.class)
+public @interface ValidRange {
 
     boolean required() default true;
+
+    String startFieldName() default "startDate";
+
+    String endFieldName() default "endDate";
 }
