+++ src/main/java/coz/weavon/validation/validator/PaginationValidator.java
@@ -1,8 +1,8 @@
-package coz.weavon.common.presentation.validator;
+package coz.weavon.validation.validator;
 
 import coz.weavon.exception.model.ClientException;
 import coz.weavon.common.presentation.model.reqeuest.PageRequest;
-import coz.weavon.common.presentation.model.validation.ValidPagination;
+import coz.weavon.validation.model.ValidPagination;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import java.util.Objects;
